added more base_urls and exported csv
