+++ biggest_stock_gainers.csv
@@ -1,0 +1,51 @@
+,symbol,price,name,change,changesPercentage,exchange
+0,BURU,0.368,"Nuburu, Inc.",0.166,82.17822,AMEX
+1,HCTI,0.0081,"Healthcare Triangle, Inc.",-0.001,-10.98901,NASDAQ
+2,NVDA,135.13,NVIDIA Corporation,-4.06,-2.91688,NASDAQ
+3,LCID,2.23,"Lucid Group, Inc.",-0.18,-7.46888,NASDAQ
+4,DNN,1.59,Denison Mines Corp.,-0.01,-0.625,AMEX
+5,SOXL,16.17,Direxion Daily Semiconductor Bull 3X Shares,-1.03,-5.98837,AMEX
+6,SLE,0.195,"Super League Enterprise, Inc.",0.038,24.20382,NASDAQ
+7,PLTR,131.78,Palantir Technologies Inc.,9.46,7.73381,NASDAQ
+8,TSLL,14.91,Direxion Daily TSLA Bull 1.5X Shares,-1.14,-7.1028,NASDAQ
+9,BMGL,6.08,Basel Medical Group Ltd Ordinary Shares,5.04,484.61538,NASDAQ
+10,PLUG,0.8827,Plug Power Inc.,-0.0437,-4.71718,NASDAQ
+11,BBAI,4.16,"BigBear.ai Holdings, Inc.",0.29,7.49354,NYSE
+12,SOXS,12.96,Direxion Daily Semiconductor Bear 3X Shares,0.75,6.14251,AMEX
+13,F,10.38,Ford Motor Company,0.16,1.56556,NYSE
+14,GNLN,0.0062,"Greenlane Holdings, Inc.",0.0,0.0,NASDAQ
+15,SQQQ,23.73,ProShares UltraPro Short QQQ,0.15,0.63613,NASDAQ
+16,TSLA,346.46,"Tesla, Inc.",-11.97,-3.33956,NASDAQ
+17,INTC,19.55,Intel Corporation,-0.7,-3.45679,NASDAQ
+18,ECDA,0.3404,"ECD Automotive Design, Inc.",0.0757,28.59841,NASDAQ
+19,TQQQ,70.0,ProShares UltraPro QQQ,-0.37,-0.52579,NASDAQ
+20,SPXS,5.45,Direxion Daily S&P 500 Bear 3X Shares,0.02,0.36832,AMEX
+21,VTRS,8.79,Viatris Inc.,-0.04,-0.453,NASDAQ
+22,SPY,589.39,SPDR S&P 500 ETF Trust,-0.66,-0.11185,AMEX
+23,BTE,1.61,Baytex Energy Corp.,-0.07,-4.16667,NYSE
+24,TSLZ,1.565,T-Rex 2X Inverse Tesla Daily Target ETF,0.105,7.19178,NASDAQ
+25,AAPL,200.85,Apple Inc.,0.9,0.45011,NASDAQ
+26,OPEN,0.6522,Opendoor Technologies Inc.,-0.0338,-4.92711,NASDAQ
+27,QBTS,16.33,D-Wave Quantum Inc.,0.06,0.36878,NYSE
+28,QQQ,519.11,Invesco QQQ Trust,-0.82,-0.15771,NASDAQ
+29,LVWR,4.5,"LiveWire Group, Inc.",1.63,56.79443,NYSE
+30,AAL,11.41,American Airlines Group Inc.,0.01,0.0877193,NASDAQ
+31,FXI,35.0,iShares China Large-Cap ETF,-0.62,-1.7406,AMEX
+32,EOSE,4.16,"Eos Energy Enterprises, Inc.",-0.355,-7.86268,NASDAQ
+33,SOFI,13.3,"SoFi Technologies, Inc.",0.13,0.98709,NASDAQ
+34,BBD,2.88,Banco Bradesco S.A.,0.0,0.0,NYSE
+35,GAP,22.31,"The Gap, Inc.",-5.64,-20.17889,NYSE
+36,GITS,2.41,"Global Interactive Technologies, Inc.",1.46,153.68421,NASDAQ
+37,NU,12.01,Nu Holdings Ltd.,0.02,0.16681,NYSE
+38,LGMK,0.0115,"LogicMark, Inc.",-0.0007,-5.7377,NASDAQ
+39,RGTI,12.11,"Rigetti Computing, Inc.",-1.04,-7.90875,NASDAQ
+40,NVDQ,1.895,T-Rex 2X Inverse NVIDIA Daily Target ETF,0.115,6.46067,AMEX
+41,AMZN,205.01,"Amazon.com, Inc.",-0.69,-0.33544,NASDAQ
+42,GOOGL,171.74,Alphabet Inc.,-0.12,-0.06982428,NASDAQ
+43,WBA,11.25,"Walgreens Boots Alliance, Inc.",0.015,0.13351,NASDAQ
+44,TLT,86.28,iShares 20+ Year Treasury Bond ETF,0.13,0.1509,NASDAQ
+45,INAB,0.1482,"IN8bio, Inc.",-0.0052,-3.38983,NASDAQ
+46,MRVL,60.19,"Marvell Technology, Inc.",-3.54,-5.55468,NASDAQ
+47,IPG,23.96,"The Interpublic Group of Companies, Inc.",-0.35,-1.43974,NYSE
+48,PATH,13.31,UiPath Inc.,0.37,2.85935,NYSE
+49,WBD,9.97,"Warner Bros. Discovery, Inc.",-0.07,-0.69721,NASDAQ
