+++ src/main.py
@@ -2,7 +2,8 @@ print("Starting main.py")
 
 def main():
   print("end of main")
-  print("it finally worked")
+  print("it finally")
+  
 if __name__=="__main__":
   main()
